removing some warnings during compilation
git-svn-id: svn+ssh://svn.forge.objectweb.org/svnroot/paje/trunk/Paje@232 c178e0d1-05e5-4583-bd9e-604f5cafc20f

+++ General/SourceCodeReference.h
@@ -44,7 +44,7 @@
 //
 
 - (NSString *)filename;
-- (int)lineNumber;
+- (NSInteger)lineNumber;
 
 //
 // it can be hashed (NSObject protocol)
